+++ tests/CreateNavigationTest.php
@@ -5,8 +5,8 @@ use Pest\Expectation;
 
 use function Pest\Laravel\assertDatabaseHas;
 
-use RyanChandler\FilamentNavigation\Filament\Resources\NavigationResource\Pages\CreateNavigation;
-use RyanChandler\FilamentNavigation\Models\Navigation;
+use Aflorea4\FilamentNavigation\Filament\Resources\NavigationResource\Pages\CreateNavigation;
+use Aflorea4\FilamentNavigation\Models\Navigation;
 
 it('can create a navigation menu', function () {
     Livewire::test(CreateNavigation::class)